CREATE: create_luas_tanam view && UPDATE: pompa and luas tanam feature

+++ app/Http/Controllers/PjKecamatan/KecamatanPompaDimanfaatkanController.php
@@ -79,6 +79,7 @@ class KecamatanPompaDimanfaatkanController extends Controller
         $user = Auth::user();
         $dimanfaatkan = PompaDimanfaatkan::find(Crypt::decryptString($id));
         if (!$dimanfaatkan) return back()->withErrors('data pompa dimanfaatkan tidak ditemukan');
+        if ($dimanfaatkan->status == 'diverifikasi') return back()->withErrors('data yang telah diverifikasi tidak dapat diubah');
         $request->validate([
             'total_unit' => 'required'
         ], [
@@ -97,6 +98,7 @@ class KecamatanPompaDimanfaatkanController extends Controller
     public function destroy($id) {
         $dimanfaatkan = PompaDimanfaatkan::find(Crypt::decryptString($id));
         if (!$dimanfaatkan) return back()->withErrors('data pompa dimanfaatkan tidak ditemukan');
+        if ($dimanfaatkan->status == 'diverifikasi') return back()->withErrors('data yang telah diverifikasi tidak dapat dihapus');
         if (!$dimanfaatkan->delete()) return back()->withErrors('terjadi kesalahan');
         return back()->with('success', 'data pompa dimanfaatkan berhasil dihapus');
     }
